Commit generated code
By doing this there is no error when composer installing or updating. This doesn't matter most of the time. But on slower hardware this is hella slow. Committing the generated code isn't a problem because on install and update it gets overwritten. So code is always update to date. (Unless someone is so "smart" to disable plugins and scripts.)